pyproject.toml
[project]
name = "mcp-bonusly"
version = "1.1.1"
description = "MCP server for Bonusly employee recognition platform"
authors = [
    { name = "Angel Ramos" }
]
dependencies = [
    "mcp>=1.0.0",
    "httpx>=0.25.0",
    "pydantic>=2.0.0",
    "email-validator>=2.0.0",
    "python-dateutil>=2.8.0",
]
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
keywords = ["mcp", "bonusly", "employee-recognition", "api"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

[project.urls]
Homepage = "https://github.com/ajramos/mcp-bonusly"
Repository = "https://github.com/ajramos/mcp-bonusly"

[project.scripts]
mcp-bonusly = "mcp_bonusly.server:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src" 